Split backwards into multiple files
