+++ src/autograd/mod.rs
@@ -1,4 +1,15 @@
+pub(super) mod util;
+pub(crate) mod into;
+
 pub mod gradient_function;
-pub mod into;
-pub mod ops_backwards;
+
+pub mod none_backwards;
+pub mod accumulate_grad;
+
+pub mod add_backwards;
+pub mod sub_backwards;
+pub mod mul_backwards;
+pub mod div_backwards;
+pub mod neg_backwards;
+
 pub mod backwards;
